[IMP] web: improve restoration of action
Before this commit, the current action was always stored, but it was
restored only if there was no action ID, and no res ID.
The issue with this, is that if that the action is modified and some
domain is injected, the domain will not be taken into account at the
reload of the page.

Now, the action is always restored.

task-id 4031939

Part-of: odoo/odoo#185674
Related: odoo/enterprise#72986
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Luca Vitali <luvi@odoo.com>
Co-authored-by: Jorge Pinna Puissant <jpp@odoo.com>